Refactor orchest-api utils
This commits moves a number of functions into the image_utils module,
namespace_info and the new environments module.
